+++ src/Core/Spec/Spec13.php
@@ -87,4 +87,9 @@ final class Spec13 implements SpecInterface
     {
         return true;
     }
+
+    public function supportsExternalReferenceHashes(): bool
+    {
+        return true;
+    }
 }