+++ tests/_data/BomModelProvider.php
@@ -24,9 +24,11 @@ declare(strict_types=1);
 namespace CycloneDX\Tests\_data;
 
 use CycloneDX\Core\Enums\Classification;
+use CycloneDX\Core\Enums\ExternalReferenceType;
 use CycloneDX\Core\Enums\HashAlgorithm;
 use CycloneDX\Core\Models\Bom;
 use CycloneDX\Core\Models\Component;
+use CycloneDX\Core\Models\ExternalReference;
 use CycloneDX\Core\Models\License\DisjunctiveLicenseWithId;
 use CycloneDX\Core\Models\License\DisjunctiveLicenseWithName;
 use CycloneDX\Core\Models\License\LicenseExpression;
@@ -34,6 +36,7 @@ use CycloneDX\Core\Models\MetaData;
 use CycloneDX\Core\Models\Tool;
 use CycloneDX\Core\Repositories\ComponentRepository;
 use CycloneDX\Core\Repositories\DisjunctiveLicenseRepository;
+use CycloneDX\Core\Repositories\ExternalReferenceRepository;
 use CycloneDX\Core\Repositories\HashRepository;
 use CycloneDX\Core\Repositories\ToolRepository;
 use Generator;
@@ -55,17 +58,49 @@ abstract class BomModelProvider
         yield from self::bomWithAllComponents();
 
         yield from self::bomWithAllMetadata();
+
+        yield from self::bomWithExternalReferences();
     }
 
     /**
-     * Just an plain BOM.
+     * Just a plain BOM.
      *
      * @psalm-return Generator<array{Bom}>
      */
     public static function bomPlain(): Generator
     {
-        yield 'plain' => [new Bom()];
-        yield 'plain v23' => [(new Bom())->setVersion(23)];
+        yield 'bom plain' => [new Bom()];
+        yield 'bom plain v23' => [(new Bom())->setVersion(23)];
+    }
+
+    /**
+     * BOM with externalReferences.
+     *
+     * @psalm-return Generator<array{Bom}>
+     */
+    public static function bomWithExternalReferences(): Generator
+    {
+        yield 'bom with empty ExternalReferences' => [
+            (new Bom())->setExternalReferenceRepository(
+                new ExternalReferenceRepository()
+            ),
+        ];
+
+        foreach (self::externalReferencesForAllTypes() as $label => $extRef) {
+            yield "bom with $label" => [
+                (new Bom())->setExternalReferenceRepository(
+                    new ExternalReferenceRepository($extRef)
+                ),
+            ];
+        }
+
+        foreach (self::externalReferencesForHashAlgorithmsAllKnown() as $label => $extRef) {
+            yield "bom with $label" => [
+                (new Bom())->setExternalReferenceRepository(
+                    new ExternalReferenceRepository($extRef)
+                ),
+            ];
+        }
     }
 
     /**
@@ -85,7 +120,7 @@ abstract class BomModelProvider
         yield from self::bomWithComponentLicenseUrl();
 
         yield from self::bomWithComponentHashAlgorithmsAllKnown();
-
+        yield from self::bomWithComponentWithExternalReferences();
         yield from self::bomWithComponentTypeAllKnown();
     }
 
@@ -133,6 +168,34 @@ abstract class BomModelProvider
             ...BomSpecData::getClassificationEnumForVersion('1.2'),
             ...BomSpecData::getClassificationEnumForVersion('1.3')
         );
+    }
+
+    /**
+     * BOM with externalReferences.
+     *
+     * @psalm-return Generator<array{Bom}>
+     */
+    public static function bomWithComponentWithExternalReferences(): Generator
+    {
+        yield 'component with empty ExternalReferences' => [
+            (new Bom())->setComponentRepository(
+                new ComponentRepository(
+                    (new Component(Classification::LIBRARY, 'dummy', 'foo-beta'))
+                        ->setExternalReferenceRepository(new ExternalReferenceRepository())
+                )
+            ),
+        ];
+
+        foreach (self::externalReferencesForAllTypes() as $label => $extRef) {
+            yield "component with $label" => [
+                (new Bom())->setComponentRepository(
+                    new ComponentRepository(
+                        (new Component(Classification::LIBRARY, 'dummy', 'foo-beta'))
+                            ->setExternalReferenceRepository(new ExternalReferenceRepository($extRef))
+                    )
+                ),
+            ];
+        }
     }
 
     /**
@@ -286,6 +349,21 @@ abstract class BomModelProvider
         }
     }
 
+    /** @psalm-return list<string> */
+    private static function allHashAlgorithms(): array
+    {
+        /** @psalm-var list<string> $known */
+        $known = array_values((new \ReflectionClass(HashAlgorithm::class))->getConstants());
+
+        return array_values(array_unique(array_merge(
+            $known,
+            BomSpecData::getHashAlgEnumForVersion('1.0'),
+            BomSpecData::getHashAlgEnumForVersion('1.1'),
+            BomSpecData::getHashAlgEnumForVersion('1.2'),
+            BomSpecData::getHashAlgEnumForVersion('1.3')
+        ), \SORT_STRING));
+    }
+
     /**
      * BOMs with all hash algorithms known.
      *
@@ -293,14 +371,8 @@ abstract class BomModelProvider
      */
     public static function bomWithComponentHashAlgorithmsAllKnown(): Generator
     {
-        /** @psalm-var list<string> $known */
-        $known = array_values((new \ReflectionClass(HashAlgorithm::class))->getConstants());
         yield from self::bomWithComponentHashAlgorithms(
-            ...$known,
-            ...BomSpecData::getHashAlgEnumForVersion('1.0'),
-            ...BomSpecData::getHashAlgEnumForVersion('1.1'),
-            ...BomSpecData::getHashAlgEnumForVersion('1.2'),
-            ...BomSpecData::getHashAlgEnumForVersion('1.3'),
+            ...self::allHashAlgorithms()
         );
     }
 
@@ -482,5 +554,39 @@ abstract class BomModelProvider
                 )
             ),
         ];
+    }
+
+    /**
+     * @return Generator<ExternalReference>
+     */
+    public static function externalReferencesForAllTypes(): Generator
+    {
+        /** @psalm-var list<string> $known */
+        $known = array_values((new \ReflectionClass(ExternalReferenceType::class))->getConstants());
+        $all = array_unique(
+            array_merge(
+                $known,
+                BomSpecData::getExternalReferenceTypeForVersion('1.1'),
+                BomSpecData::getExternalReferenceTypeForVersion('1.2'),
+                BomSpecData::getExternalReferenceTypeForVersion('1.3'),
+            )
+        );
+
+        foreach ($all as $type) {
+            yield "externalReferenceType: $type" => new ExternalReference($type, ".../types/{$type}.txt");
+        }
+    }
+
+    /**
+     * BOMs with all hash algorithms known.
+     *
+     * @psalm-return Generator<array{Bom}>
+     */
+    public static function externalReferencesForHashAlgorithmsAllKnown(): Generator
+    {
+        $type = ExternalReferenceType::OTHER;
+        foreach (self::allHashAlgorithms() as $algorithm) {
+            yield "externalReferenceHash: $algorithm" => (new ExternalReference($type, ".../algorithm/{$algorithm}.txt"))->setHashRepository(new HashRepository([$algorithm => '12345678901234567890123456789012']));
+        }
     }
 }